Refactoring payment methods
